split parsers into more files and write basic documentation
