+++ docs/key-server-client.js.html
@@ -27,7 +27,7 @@
 
 <nav >
     
-    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="KeyChain.html">KeyChain</a><ul class='methods'><li data-type='method'><a href="KeyChain.html#.getKeyByField">getKeyByField</a></li><li data-type='method'><a href="KeyChain.html#.getKeyBySubKeyId">getKeyBySubKeyId</a></li><li data-type='method'><a href="KeyChain.html#.getSubKeyIdByCapability">getSubKeyIdByCapability</a></li><li data-type='method'><a href="KeyChain.html#.isKeyFromCard">isKeyFromCard</a></li><li data-type='method'><a href="KeyChain.html#call">call</a></li><li data-type='method'><a href="KeyChain.html#cardStatus">cardStatus</a></li><li data-type='method'><a href="KeyChain.html#decrypt">decrypt</a></li><li data-type='method'><a href="KeyChain.html#encrypt">encrypt</a></li><li data-type='method'><a href="KeyChain.html#generateKey">generateKey</a></li><li data-type='method'><a href="KeyChain.html#hasCard">hasCard</a></li><li data-type='method'><a href="KeyChain.html#isCardTrusted">isCardTrusted</a></li><li data-type='method'><a href="KeyChain.html#listPublicKeys">listPublicKeys</a></li><li data-type='method'><a href="KeyChain.html#listSecretKeys">listSecretKeys</a></li><li data-type='method'><a href="KeyChain.html#lookupKey">lookupKey</a></li><li data-type='method'><a href="KeyChain.html#open">open</a></li><li data-type='method'><a href="KeyChain.html#recvKey">recvKey</a></li><li data-type='method'><a href="KeyChain.html#refreshKeys">refreshKeys</a></li><li data-type='method'><a href="KeyChain.html#sendKeys">sendKeys</a></li><li data-type='method'><a href="KeyChain.html#signKey">signKey</a></li><li data-type='method'><a href="KeyChain.html#tar">tar</a></li><li data-type='method'><a href="KeyChain.html#trustCard">trustCard</a></li><li data-type='method'><a href="KeyChain.html#trustKey">trustKey</a></li><li data-type='method'><a href="KeyChain.html#verify">verify</a></li><li data-type='method'><a href="KeyChain.html#whoami">whoami</a></li></ul></li><li><a href="KeyServerClient.html">KeyServerClient</a><ul class='methods'><li data-type='method'><a href="KeyServerClient.html#fetch">fetch</a></li><li data-type='method'><a href="KeyServerClient.html#search">search</a></li></ul></li></ul><h3><a href="global.html">Global</a></h3>
+    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="GpgParser.html">GpgParser</a><ul class='methods'><li data-type='method'><a href="GpgParser.html#.parseColons">parseColons</a></li><li data-type='method'><a href="GpgParser.html#.parseStatusFd">parseStatusFd</a></li><li data-type='method'><a href="GpgParser.html#.Status_AssertSignatureAllowedstatic">Status_AssertSignatureAllowedstatic</a></li><li data-type='method'><a href="GpgParser.html#parseReaderColons">parseReaderColons</a></li><li data-type='method'><a href="GpgParser.html#Status_GetImportedKeys">Status_GetImportedKeys</a></li><li data-type='method'><a href="GpgParser.html#Status_GetSigPrimaryFpr">Status_GetSigPrimaryFpr</a></li><li data-type='method'><a href="GpgParser.html#Status_GetSigResult">Status_GetSigResult</a></li><li data-type='method'><a href="GpgParser.html#Status_GetSigTrustResult">Status_GetSigTrustResult</a></li><li data-type='method'><a href="GpgParser.html#Status_HasKeyword">Status_HasKeyword</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigBad">Status_IsSigBad</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigError">Status_IsSigError</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigExpired">Status_IsSigExpired</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigGood">Status_IsSigGood</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigKeyExpired">Status_IsSigKeyExpired</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigKeyRevoked">Status_IsSigKeyRevoked</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSignerAllowed">Status_IsSignerAllowed</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigTrustFully">Status_IsSigTrustFully</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigTrustMarginal">Status_IsSigTrustMarginal</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigTrustNever">Status_IsSigTrustNever</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigTrustUltimate">Status_IsSigTrustUltimate</a></li><li data-type='method'><a href="GpgParser.html#Status_IsSigTrustUnknown">Status_IsSigTrustUnknown</a></li></ul></li><li><a href="KeyChain.html">KeyChain</a><ul class='methods'><li data-type='method'><a href="KeyChain.html#.getKeyByField">getKeyByField</a></li><li data-type='method'><a href="KeyChain.html#.getKeyBySubKeyId">getKeyBySubKeyId</a></li><li data-type='method'><a href="KeyChain.html#.getSubKeyIdByCapability">getSubKeyIdByCapability</a></li><li data-type='method'><a href="KeyChain.html#.isKeyFromCard">isKeyFromCard</a></li><li data-type='method'><a href="KeyChain.html#call">call</a></li><li data-type='method'><a href="KeyChain.html#cardStatus">cardStatus</a></li><li data-type='method'><a href="KeyChain.html#decrypt">decrypt</a></li><li data-type='method'><a href="KeyChain.html#encrypt">encrypt</a></li><li data-type='method'><a href="KeyChain.html#generateKey">generateKey</a></li><li data-type='method'><a href="KeyChain.html#hasCard">hasCard</a></li><li data-type='method'><a href="KeyChain.html#isCardTrusted">isCardTrusted</a></li><li data-type='method'><a href="KeyChain.html#listPublicKeys">listPublicKeys</a></li><li data-type='method'><a href="KeyChain.html#listSecretKeys">listSecretKeys</a></li><li data-type='method'><a href="KeyChain.html#lookupKey">lookupKey</a></li><li data-type='method'><a href="KeyChain.html#open">open</a></li><li data-type='method'><a href="KeyChain.html#recvKey">recvKey</a></li><li data-type='method'><a href="KeyChain.html#refreshKeys">refreshKeys</a></li><li data-type='method'><a href="KeyChain.html#sendKeys">sendKeys</a></li><li data-type='method'><a href="KeyChain.html#signKey">signKey</a></li><li data-type='method'><a href="KeyChain.html#tar">tar</a></li><li data-type='method'><a href="KeyChain.html#trustCard">trustCard</a></li><li data-type='method'><a href="KeyChain.html#trustKey">trustKey</a></li><li data-type='method'><a href="KeyChain.html#verify">verify</a></li><li data-type='method'><a href="KeyChain.html#whoami">whoami</a></li></ul></li><li><a href="KeyServerClient.html">KeyServerClient</a><ul class='methods'><li data-type='method'><a href="KeyServerClient.html#fetch">fetch</a></li><li data-type='method'><a href="KeyServerClient.html#search">search</a></li></ul></li></ul><h3><a href="global.html">Global</a></h3>
 </nav>
 
 <div id="main">
@@ -185,7 +185,7 @@ module.exports = KeyServerClient</code></pre>
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Wed Jan 15 2020 03:21:43 GMT-0800 (Pacific Standard Time) using the <a href="https://github.com/clenemt/docdash">docdash</a> theme.
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Tue Feb 18 2020 22:37:35 GMT-0800 (Pacific Standard Time) using the <a href="https://github.com/clenemt/docdash">docdash</a> theme.
 </footer>
 
 <script>prettyPrint();</script>
