+++ docs/styles/jsdoc.css
@@ -1,5 +1,3 @@
-@import url(https://fonts.googleapis.com/css?family=Montserrat:400,700);
-
 * {
   box-sizing: border-box
 }
@@ -119,6 +117,51 @@ h6 {
   font-style: italic;
 }
 
+.usertext h1 {
+  font-family: "Source Sans Pro";
+  font-size: 24px;
+  margin: 2.5em 0 1em;
+  font-weight: 400;
+}
+
+.usertext h2 {
+  font-family: "Source Sans Pro";
+  font-size: 18px;
+  margin: 2em 0 0.5em;
+  font-weight: 400;
+
+}
+
+.usertext h3 {
+  font-family: "Source Sans Pro";
+  font-size: 15px;
+  margin: 1.5em 0 0;
+  font-weight: 400;
+}
+
+.usertext h4 {
+  font-family: "Source Sans Pro";
+  font-size: 14px;
+  margin: 0 0 0;
+  font-weight: 400;
+}
+
+.usertext h5 {
+  font-size: 12px;
+  margin: 1em 0 0;
+  font-weight: normal;
+  color: #666;
+}
+
+.usertext h6 {
+  font-size: 11px;
+  margin: 1em 0 0;
+  font-weight: normal;
+  font-style: normal;
+  color: #666;
+}
+
+
 tt, code, kbd, samp {
   font-family: Consolas, Monaco, 'Andale Mono', monospace;
   background: #f4f4f4;
@@ -138,7 +181,6 @@ tt, code, kbd, samp {
 
 #main {
   float: right;
-  min-width: 360px;
   width: calc(100% - 240px);
 }
 
@@ -408,6 +450,12 @@ footer {
    user-select: text;
 }
 
+.prettyprint.linenums li code:empty:after {
+  content:"";
+  display:inline-block;
+  width:0px;
+}
+
 table {
   border-spacing: 0;
   border: 1px solid #ddd;
@@ -613,7 +661,7 @@ html[data-search-mode] .level-hide {
 }
 
 
-@media only screen and (min-width: 320px) and (max-width: 680px) {
+@media only screen and (max-width: 680px) {
   body {
     overflow-x: hidden;
   }
@@ -642,7 +690,6 @@ html[data-search-mode] .level-hide {
 
   #main {
     width: 100%;
-    min-width: 360px;
   }
 
   #main h1.page-title {
@@ -668,4 +715,51 @@ html[data-search-mode] .level-hide {
 
 #disqus_thread{
     margin-left: 30px;
+}
+
+@font-face {
+  font-family: 'Montserrat';
+  font-style: normal;
+  font-weight: 400;
+  src: url('../fonts/Montserrat/Montserrat-Regular.eot'); /* IE9 Compat Modes */
+  src: url('../fonts/Montserrat/Montserrat-Regular.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+       url('../fonts/Montserrat/Montserrat-Regular.woff2') format('woff2'), /* Super Modern Browsers */
+       url('../fonts/Montserrat/Montserrat-Regular.woff') format('woff'), /* Pretty Modern Browsers */
+       url('../fonts/Montserrat/Montserrat-Regular.ttf')  format('truetype'); /* Safari, Android, iOS */
+}
+
+@font-face {
+  font-family: 'Montserrat';
+  font-style: normal;
+  font-weight: 700;
+  src: url('../fonts/Montserrat/Montserrat-Bold.eot'); /* IE9 Compat Modes */
+  src: url('../fonts/Montserrat/Montserrat-Bold.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+       url('../fonts/Montserrat/Montserrat-Bold.woff2') format('woff2'), /* Super Modern Browsers */
+       url('../fonts/Montserrat/Montserrat-Bold.woff') format('woff'), /* Pretty Modern Browsers */
+       url('../fonts/Montserrat/Montserrat-Bold.ttf')  format('truetype'); /* Safari, Android, iOS */
+}
+
+@font-face {
+    font-family: 'Source Sans Pro';
+    src: url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.eot');
+    src: url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.eot?#iefix') format('embedded-opentype'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.woff2') format('woff2'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.woff') format('woff'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.ttf') format('truetype'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-regular-webfont.svg#source_sans_proregular') format('svg');
+    font-weight: 400;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Source Sans Pro';
+    src: url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.eot');
+    src: url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.eot?#iefix') format('embedded-opentype'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.woff2') format('woff2'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.woff') format('woff'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.ttf') format('truetype'),
+         url('../fonts/Source-Sans-Pro/sourcesanspro-light-webfont.svg#source_sans_prolight') format('svg');
+    font-weight: 300;
+    font-style: normal;
+
 }
